requirements.txt
asgiref==3.7.2
Django==4.2.6
django-cors-headers==4.2.0
django-unfold==0.43.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
openpyxl==3.1.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2024.1
